New folder/requirements.txt
absl-py
annotated-types
anyio
astunparse
cachetools
certifi
charset-normalizer
click
colorama
ecdsa
exceptiongroup
fastapi
flatbuffers
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
httplib2
idna
joblib
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
namex
numpy
opt-einsum
optree
packaging
pandas
passlib
pillow
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
requests
rich
rsa
scikit-learn
scipy
six
sniffio
SQLAlchemy
starlette
tensorboard
tensorboard-data-server
tensorflow
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tqdm
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
Werkzeug
wrapt
